Move all alchemistry functionality into the single class AlchemicalFreeEnergy
