+++ python/BioSimSpace/Sandpit/Exscientia/FreeEnergy/__init__.py
@@ -28,8 +28,7 @@ Classes
 .. autosummary::
     :toctree: generated/
 
-    Absolute
-    Relative
+    AlchemicalFreeEnergy
     Restraint
     RestraintSearch
 
@@ -43,8 +42,7 @@ Functions
     getData
 """
 
-from ._absolute import *
-from ._relative import *
+from ._alchemical_free_energy import *
 from ._utils import *
 from ._restraint import *
 from ._restraint_search import *
