+++ tests/Sandpit/Exscientia/FreeEnergy/test_alchemical_free_energy.py
@@ -1,3 +1,11 @@
+import bz2
+import pandas as pd
+import pathlib
+import pytest
+import numpy as np
+import os
+import time
+
 try:
     from alchemlyb.parsing.gmx import extract_u_nk
 
@@ -5,14 +13,16 @@ try:
 except ModuleNotFoundError:
     is_alchemlyb = False
 
-import numpy as np
-import os
-import pandas as pd
-import pathlib
-import pytest
+try:
+    from alchemtest.gmx import load_ABFE
+    from alchemtest.amber import load_bace_example
 
-import time
+    is_alchemtest = True
+except ModuleNotFoundError:
+    is_alchemtest = False
 
+import BioSimSpace.Sandpit.Exscientia as BSS
+from BioSimSpace.Sandpit.Exscientia import Types as _Types
 from BioSimSpace.Sandpit.Exscientia.Align._decouple import decouple
 from BioSimSpace.Sandpit.Exscientia.FreeEnergy import Restraint
 from BioSimSpace.Sandpit.Exscientia.Protocol import FreeEnergyEquilibration
@@ -20,16 +30,105 @@ from BioSimSpace.Sandpit.Exscientia.Units.Angle import radian, degree
 from BioSimSpace.Sandpit.Exscientia.Units.Energy import kcal_per_mol
 from BioSimSpace.Sandpit.Exscientia.Units.Length import angstrom
 from BioSimSpace.Sandpit.Exscientia.Units.Temperature import kelvin
-from BioSimSpace.Sandpit.Exscientia import Types as _Types
-import BioSimSpace.Sandpit.Exscientia as BSS
 
-
-# Make sure GROMSCS is installed.
+# Make sure GROMACS is installed.
 has_gromacs = BSS._gmx_exe is not None
 
-# Get the input files
+# Store the tutorial URL.
 url = BSS.tutorialUrl()
 
+
+@pytest.mark.skipif(
+    is_alchemtest is False, reason="Requires alchemtest and alchemlyb to be installed."
+)
+class TestRelativeAnalysis:
+    @staticmethod
+    @pytest.fixture(scope="class")
+    def gmx_data(tmp_path_factory):
+        outdir = tmp_path_factory.mktemp("gromacs")
+        for leg in ["complex", "ligand"]:
+            for index, filename in enumerate(load_ABFE().data[leg]):
+                dir = outdir / leg / f"lambda_{index}"
+                dir.mkdir(parents=True)
+                eng = dir / "gromacs.xvg"
+                eng.symlink_to(filename)
+        return outdir
+
+    @staticmethod
+    @pytest.fixture(scope="class")
+    def amber_data(tmp_path_factory):
+        outdir = tmp_path_factory.mktemp("amber")
+        for leg, stage in [("complex", "decharge"), ("solvated", "vdw")]:
+            for index, filename in enumerate(load_bace_example().data[leg][stage]):
+                dir = outdir / f"{leg}_{stage}" / f"lambda_{index}"
+                dir.mkdir(parents=True)
+                with open(dir / "amber.out", "w") as f:
+                    with bz2.open(filename, "rt") as bz_file:
+                        f.write(bz_file.read())
+        return outdir
+
+    @staticmethod
+    @pytest.fixture(scope="class")
+    def gmx_complex(gmx_data):
+        outdir = gmx_data
+        complex, _ = BSS.FreeEnergy.AlchemicalFreeEnergy.analyse(
+            work_dir=str(outdir / "complex"),
+            temperature=310 * BSS.Units.Temperature.kelvin,
+        )
+        return complex
+
+    @staticmethod
+    @pytest.fixture(scope="class")
+    def gmx_ligand(gmx_data):
+        outdir = gmx_data
+        ligand, _ = BSS.FreeEnergy.AlchemicalFreeEnergy.analyse(
+            work_dir=str(outdir / "ligand"),
+            temperature=310 * BSS.Units.Temperature.kelvin,
+        )
+        return ligand
+
+    @staticmethod
+    @pytest.fixture(scope="class")
+    def amber_complex_decharge(amber_data):
+        outdir = amber_data
+        complex, _ = BSS.FreeEnergy.AlchemicalFreeEnergy.analyse(
+            work_dir=str(outdir / "complex_decharge"),
+            temperature=298 * BSS.Units.Temperature.kelvin,
+        )
+        return complex
+
+    @staticmethod
+    @pytest.fixture(scope="class")
+    def amber_solvated_vdw(amber_data):
+        outdir = amber_data
+        complex, _ = BSS.FreeEnergy.AlchemicalFreeEnergy.analyse(
+            work_dir=str(outdir / "solvated_vdw"),
+            temperature=298 * BSS.Units.Temperature.kelvin,
+        )
+        return complex
+
+    @pytest.mark.parametrize(
+        "fixture,length,energy",
+        [
+            ("gmx_ligand", 20, 7.654472744451637),
+            ("gmx_complex", 30, 21.819752),
+            ("amber_complex_decharge", 5, -5.25352),
+            ("amber_solvated_vdw", 12, 2.261816),
+        ],
+    )
+    def test_pmf(self, fixture, length, energy, request):
+        pmf = request.getfixturevalue(fixture)
+        assert len(pmf) == length
+        assert len(pmf[0]) == 3
+        np.testing.assert_allclose(
+            pmf[-1][1] / BSS.Units.Energy.kcal_per_mol, energy, atol=0.1
+        )
+
+    def test_difference(self, gmx_complex, gmx_ligand):
+        dG, error = BSS.FreeEnergy.AlchemicalFreeEnergy.difference(gmx_complex, gmx_ligand)
+        np.testing.assert_allclose(
+            dG / BSS.Units.Energy.kcal_per_mol, 14.216101, atol=0.1
+        )
 
 @pytest.mark.skipif(has_gromacs is False, reason="Requires GROMACS to be installed.")
 @pytest.mark.skipif(is_alchemlyb is False, reason="Requires alchemlyb to be installed.")
@@ -100,7 +199,7 @@ class Test_gmx_ABFE:
             ),
             runtime=_Types.Time(0, "nanoseconds"),
         )
-        freenrg = BSS.FreeEnergy.Absolute(
+        freenrg = BSS.FreeEnergy.AlchemicalFreeEnergy(
             solvated, protocol, engine="GROMACS", restraint=restraint
         )
         freenrg.run()
@@ -183,7 +282,7 @@ class Test_Somd_ABFE:
         protocol = BSS.Protocol.FreeEnergy(
             lam_vals=[0.0, 0.5, 1.0], runtime=0.0001 * BSS.Units.Time.nanosecond
         )
-        freenrg = BSS.FreeEnergy.Absolute(
+        freenrg = BSS.FreeEnergy.AlchemicalFreeEnergy(
             solvated, protocol, engine="SOMD", restraint=restraint
         )
 
@@ -207,7 +306,6 @@ class Test_Somd_ABFE:
             assert (path / f"lambda_{lam}" / "somd.err").is_file()
             assert (path / f"lambda_{lam}" / "somd.out").is_file()
 
-    # @pytest.mark.xfail(reason="freenrg.wait() doesn't work for SOMD, so files aren't created in time.")
     def test_correct_conf_file(self, freenrg):
         """Check that lambda data is correct in somd.cfg"""
         path = pathlib.Path(freenrg.workDir())
